refactor: extract "private" out-of-class method "_notify"
