+++ compass-js-frontend/src/main/webapp/ui/view/editor/hierarchy.js
@@ -10,6 +10,7 @@
 var $ = global.jQuery;
 var basicContext = global.basicContext;
 var AmpersandView = require('ampersand-view');
+var _notify = require('./_notify');
 var template = require('../../templates/editor/hierarchy.html');
 
 var HierarchyView = AmpersandView.extend({
@@ -120,13 +121,12 @@ var HierarchyView = AmpersandView.extend({
 			sceneNode = this.parent.root;
 		}
 		var newNode = sceneNode.children.add({name: "New Node", parentNode: sceneNode});
-		var view = this;
 		newNode.save().then(function(){
 			// success
 			// we just expect that, so no need to tell the user anything.
 		}, function() {
 			// fail
-			view.notify("error", "Could not save your new node to the server.");
+			_notify("danger", "Could not save your new node to the server.");
 		})
 		this.insertNodeIntoTree(newNode, sceneNode);
 		this.parent.selectedNode = newNode;
@@ -153,31 +153,6 @@ var HierarchyView = AmpersandView.extend({
 		//      (ideally, the button would be disabled via a binding)
 		//      complain.
 	},
-	notify: function(type, message) {
-		var options = {}
-		switch (type) {
-			case "error":
-				options.icon = "fa fa-exclamation-triangle";
-				break;
-			case "success":
-				options.icon = "fa fa-check-square-o"
-				break;
-		}
-		options.message = message;
-		var settings = {
-			type: type,
-			placement: {
-				from: "top",
-				align: "center"
-			},
-			animate: {
-				enter: 'animated fadeInDown',
-				exit: 'animated fadeOutUp'
-			},
-			allow_dismiss: false
-		};
-		$.notify(options, settings);
-	}
 });
 
 module.exports = HierarchyView;
